+++ src/App.jsx
@@ -1,10 +1,11 @@
 import React from "react";
-import TodoList from "./components/TodoList";
+
+import TodoContainer from "./components/TodoContainer";
 
 const App = () => {
   return (
     <main className="main-center">
-      <TodoList />
+      <TodoContainer />
     </main>
   );
 };
