+++ src/constants/sample-todo.js
@@ -1,0 +1,12 @@
+export const SAMPLE_TODOS = [
+  { id: 1, text: "Buy milk", completed: false },
+  { id: 2, text: "Clean the house", completed: false },
+  { id: 3, text: "Go for a run", completed: false },
+  { id: 4, text: "Finish homework", completed: false },
+  { id: 5, text: "Call mom", completed: false },
+  { id: 6, text: "Buy groceries", completed: false },
+  { id: 7, text: "Walk the dog", completed: false },
+  { id: 8, text: "Read a book", completed: false },
+  { id: 9, text: "Do laundry", completed: false },
+  { id: 10, text: "Write code", completed: false },
+];
